Add some helpers and start adding react-table
