+++ package.json
@@ -67,6 +67,7 @@
     "react-lazyload": "^3.2.0",
     "react-markdown": "^5.0.3",
     "react-redux": "^7.2.6",
+    "react-table": "^7.7.0",
     "react-tooltip": "^4.2.21",
     "react-transition-group": "^2.0.1",
     "react-use-intercom": "^1.2.0",
