+++ src/ui/components/NetworkMonitor/Status.module.css
@@ -1,3 +1,7 @@
+.status {
+  line-height: 16px;
+}
+
 .failure {
   background: var(--red-50);
   color: white;
